Merged PR 4083: Update method of parsing multipart bodies
Update method of parsing multipart bodies

The previous (simpler) way did not consistently work. Seemingly randomly it would not parse all the body's fields. The new implementation (while a bit less simple) does seem to work consistently.
However, it does assume the name of the field that holds the file's content is 'file'. We can parameterize/refactor this when needed.

Related work items: #7578